Consolidate expression and literal tests
